requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
numpy
python-dotenv
motor
certifi
